+++ src/aws_cost_optimizer/__init__.py
@@ -1,5 +1,5 @@
 """AWS Cost Optimizer - Automated cost reduction recommendations"""
 
-__version__ = "0.1.0"
+__version__ = "0.1.3"
 __author__ = "Jonathan Schimpf"
 __email__ = "jon@theatrico.org"